+++ src/container/modules/node-transformers/FinalizingTransformersModule.ts
@@ -1,0 +1,6 @@
+import { ContainerModule, interfaces } from 'inversify';
+
+export const finalizingTransformersModule: interfaces.ContainerModule = new ContainerModule((bind: interfaces.Bind) => {
+    // finalizing transformers
+
+});