+++ src/interfaces/source-map/ISourceMapCorrector.d.ts
@@ -2,9 +2,9 @@ import { IObfuscationResult } from '../IObfuscationResult';
 
 export interface ISourceMapCorrector {
     /**
-     * @param obfuscatedCode
-     * @param sourceMap
-     * @returns IObfuscationResult
+     * @param {string} obfuscatedCode
+     * @param {string} sourceMap
+     * @returns {IObfuscationResult}
      */
     correct (obfuscatedCode: string, sourceMap: string): IObfuscationResult;
 }