+++ test/functional-tests/node-transformers/preparing-transformers/eval-call-expression-transformer/fixtures/identifier-reference.js
@@ -1,0 +1,7 @@
+(function(){
+    function foo (bar) {
+        eval('bar');
+    }
+
+    foo(1);
+})();