+++ test/functional-tests/node-transformers/preparing-transformers/eval-call-expression-transformer/fixtures/multiple-statements-eval.js
@@ -1,0 +1,7 @@
+(function(){
+    function foo (bar, baz) {
+        eval('bar;baz;');
+    }
+
+    foo(1, 2);
+})();