+++ test/functional-tests/node-transformers/preparing-transformers/eval-call-expression-transformer/fixtures/wrong-eval-string.js
@@ -1,0 +1,7 @@
+(function(){
+    function foo () {
+        eval('~');
+    }
+
+    foo();
+})();